Adding display of submenu level
Caveat is that twitter bootstrap 3.0 removed multi-level support for drop down menus so I've implemented a css to make it happen, and with adjustments it should be able to work on click instead of on hover as it is now. But I've realised that multi-level menu support is probably not the best. For huge sites probably is better to use dropdown panels...